# specs/001-b2b2c/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: 商家入驻管理 API
  description: B2B2C 平台商家入驻系统 RESTful API 规范
  version: 1.0.0
  contact:
    name: API Support
    email: api@grainadmin.com
servers:
  - url: https://api.grainadmin.com/v1
    description: Production
  - url: http://localhost:8000/v1
    description: Development
security:
  - BearerAuth: []
...
paths:
  /applications:
    post:
      tags: [applications]
      summary: 创建商家入驻申请
      description: 商家创建新的入驻申请（初始状态为 DRAFT）
      operationId: createApplication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateApplicationRequest'
      responses:
        '201':
          description: 申请创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
      security:
        - BearerAuth: []

    get:
      tags: [applications]
      summary: 查询商家申请列表
      description: |
        查询商家申请列表（支持分页和筛选）
        - 商家用户：只能查询自己的申请
        - 平台管理员：可查询所有申请
      operationId: listApplications
      parameters:
        - name: status
          in: query
          description: 申请状态筛选
          schema:
            type: string
            enum: [draft, pending, under_review, require_supplement, approved, rejected]
        - name: page
          in: query
          description: 页码（从 1 开始）
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: page_size
          in: query
          description: 每页数量
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /applications/{application_id}:
    get:
      tags: [applications]
      summary: 查询申请详情
      description: |
        查询指定申请的详细信息
        - 商家用户：只能查询自己的申请
        - 平台管理员：可查询所有申请
      operationId: getApplication
      parameters:
        - $ref: '#/components/parameters/ApplicationId'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationDetailResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    put:
      tags: [applications]
      summary: 更新申请信息
      description: |
        更新商家申请信息（仅 DRAFT 和 REQUIRE_SUPPLEMENT 状态可修改）
      operationId: updateApplication
      parameters:
        - $ref: '#/components/parameters/ApplicationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateApplicationRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: 申请状态不允许修改
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /applications/{application_id}/submit:
    post:
      tags: [applications]
      summary: 提交申请
      description: |
        将申请从 DRAFT/REQUIRE_SUPPLEMENT 状态提交至 PENDING 状态
      operationId: submitApplication
      parameters:
        - $ref: '#/components/parameters/ApplicationId'
      responses:
        '200':
          description: 提交成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationResponse'
        '400':
          description: 申请信息不完整或状态不允许提交
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /audits:
    post:
      tags: [audits]
      summary: 创建审核记录
      description: |
        审核员对商家申请进行审核操作（通过/拒绝/要求补充材料）
      operationId: createAudit
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAuditRequest'
      responses:
        '201':
          description: 审核记录创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /audits/history/{application_id}:
    get:
      tags: [audits]
      summary: 查询审核历史
      description: |
        查询指定申请的完整审核历史记录
      operationId: getAuditHistory
      parameters:
        - $ref: '#/components/parameters/ApplicationId'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditHistoryResponse'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /files/upload:
    post:
      tags: [files]
      summary: 上传文件
      description: |
        上传资质文件或合同照片
      operationId: uploadFile
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: 文件内容
                file_type:
                  type: string
                  enum: [business_license, id_card_front, id_card_back, contract, other]
                  description: 文件类型
                related_id:
                  type: string
                  format: uuid
                  description: 关联 ID（application_id 或 merchant_id）
              required: [file, file_type, related_id]
      responses:
        '201':
          description: 上传成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileUploadResponse'
        '400':
          description: 文件格式或大小不符合要求
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /files/{file_id}:
    get:
      tags: [files]
      summary: 下载文件
      description: |
        下载文件（返回签名 URL 或文件流）
      operationId: downloadFile
      parameters:
        - name: file_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 文件下载成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
                    format: uri
                    description: 签名 URL（有效期 1 小时）
        '404':
          $ref: '#/components/responses/NotFound'
        '403':
          $ref: '#/components/responses/Forbidden'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ApplicationId:
      name: application_id
      in: path
      required: true
      description: 申请 ID
      schema:
        type: string
        format: uuid
  schemas:
    CreateApplicationRequest:
      type: object
      required:
        - business_name
        - unified_social_credit
        - legal_person_name
        - legal_person_id_number
        - contact_name
        - contact_phone
        - contact_email
        - business_categories
      properties:
        business_name:
          type: string
          maxLength: 200
          example: "测试科技有限公司"
        unified_social_credit:
          type: string
          pattern: '^[0-9A-HJ-NPQRTUWXY]{2}\d{6}[0-9A-HJ-NPQRTUWXY]{10}$'
          example: "91110108MA01234567"
        legal_person_name:
          type: string
          maxLength: 100
          example: "张三"
        legal_person_id_number:
          type: string
          pattern: '^\d{17}[\dXx]$'
          example: "110101199001011234"
        contact_name:
          type: string
          maxLength: 100
          example: "李四"
        contact_phone:
          type: string
          pattern: '^1[3-9]\d{9}$'
          example: "13800138000"
        contact_email:
          type: string
          format: email
          example: "contact@example.com"
        business_categories:
          type: array
          minItems: 1
          items:
            type: string
          example: ["食品", "日用品"]
    UpdateApplicationRequest:
      type: object
      properties:
        contact_name:
          type: string
          maxLength: 100
        contact_phone:
          type: string
          pattern: '^1[3-9]\d{9}$'
        contact_email:
          type: string
          format: email
        business_categories:
          type: array
          items:
            type: string
    ApplicationResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        business_name:
          type: string
        unified_social_credit:
          type: string
        legal_person_name:
          type: string
        contact_name:
          type: string
        contact_phone:
          type: string
        contact_email:
          type: string
        business_categories:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [draft, pending, under_review, require_supplement, approved, rejected]
        submitted_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ApplicationDetailResponse:
      allOf:
        - $ref: '#/components/schemas/ApplicationResponse'
        - type: object
          properties:
            qualification_files:
              type: array
              items:
                $ref: '#/components/schemas/FileInfo'
            audit_records:
              type: array
              items:
                $ref: '#/components/schemas/AuditRecordSummary'
    ApplicationListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ApplicationResponse'
        total:
          type: integer
          description: 总记录数
        page:
          type: integer
          description: 当前页码
        page_size:
          type: integer
          description: 每页数量
    CreateAuditRequest:
      type: object
      required:
        - application_id
        - action
        - result
      properties:
        application_id:
          type: string
          format: uuid
        action:
          type: string
          enum: [start_review, approve, reject, request_supplement]
        result:
          type: string
          enum: [approved, rejected, pending_supplement, in_progress]
        comment:
          type: string
          description: 审核意见（reject 和 request_supplement 时必填）
    AuditResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        application_id:
          type: string
          format: uuid
        auditor_id:
          type: string
          format: uuid
        auditor_name:
          type: string
        action:
          type: string
        result:
          type: string
        comment:
          type: string
        created_at:
          type: string
          format: date-time
    AuditRecordSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        auditor_name:
          type: string
        action:
          type: string
        result:
          type: string
        created_at:
          type: string
          format: date-time
    AuditHistoryResponse:
      type: object
      properties:
        application_id:
          type: string
          format: uuid
        records:
          type: array
          items:
            $ref: '#/components/schemas/AuditResponse'
...
    FileUploadResponse:
      type: object
      properties:
        file_id:
          type: string
          format: uuid
        file_path:
          type: string
        file_size:
          type: integer
        file_type:
          type: string
        uploaded_at:
          type: string
          format: date-time
    FileInfo:
      type: object
      properties:
        id:
          type: string
          format: uuid
        file_type:
          type: string
        file_size:
          type: integer
        audit_status:
          type: string
          enum: [pending, approved, rejected]
        uploaded_at:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: 错误类型
        message:
          type: string
          description: 错误详细信息
        details:
          type: object
          description: 额外错误详情
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: 未认证
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: 无权限访问
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'